requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==18.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.20
langchain-core==0.3.45
langchain-ollama==0.2.3
langchain-text-splitters==0.3.6
langsmith==0.3.15
MarkupSafe==3.0.2
ollama==0.4.7
orjson==3.10.15
packaging==24.2
pdflatex==0.1.3
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.39
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
zstandard==0.23.0
